requirements.txt
Django>=4.1,<4.2
djangorestframework>=3.13.1,<3.14
psycopg2>=2.9.3,<2.10
drf-spectacular>=0.23.1,<0.24
django-rest-swagger>=2.2.0,<2.3
